Move stats stuff to its own package
